refactor(extension): extract tx collateral computation to a hook
